Move into fiware folder and cleanup
Remove unnecessary comments
Remove return values and add exceptions
